Add @no1d 's bug fixes and divide the original ASSSubtitle into files by class.
