+++ ASSLoader.NET/ASSSubtitle.cs
@@ -1,3 +1,4 @@
+using ASSLoader.NET.Enums;
 using ASSLoader.NET.Exceptions;
 using System;
 using System.Collections.Generic;
@@ -11,17 +12,33 @@ namespace ASSLoader.NET
 {
     public class ASSSubtitle
     {
-        public const bool showAbstract = true;
         public Dictionary<string, Tuple<string, string>> ScriptInfo { get; set; }
+
         public IList<string> V4pStyleFormat { get; set; }
+
         public Dictionary<string, ASSStyle> V4pStyles { get; set; }
+
         public IList<string> EventFormat { get; set; }
+
         public IList<ASSEvent> Events { get; set; }
+
+        /// <summary>
+        /// *** NOT IMPLEMENT FOR NOW.
+        /// </summary>
         public Dictionary<string, ASSEmbeddedFont> Fonts { get; set; }
+
+        /// <summary>
+        /// *** NOT IMPLEMENT FOR NOW.
+        /// </summary>
         public Dictionary<string, ASSEmbeddedGraphics> Graphics { get; set; }
 
         public Dictionary<string, string> UnknownSections { get; set; }
 
+        /// <summary>
+        /// Load a ".ass" file, and generate the ASSSubtitle object.
+        /// </summary>
+        /// <param name="path">The path of the ".ass" file.</param>
+        /// <param name="enc">The encoding used to process the file. Default encoding is 'utf-8'(no BOM).</param>
         public void Load(string path, Encoding enc = null)
         {
             string[] lines;
@@ -36,9 +53,10 @@ namespace ASSLoader.NET
 
             var workingSection = ASSSection.ScriptInfo;
             var scriptInfoCommentIndex = 0;
+            var unknowSectionName = string.Empty;
+
             V4pStyleFormat = new List<string>();
             EventFormat = new List<string>();
-
             ScriptInfo = new Dictionary<string, Tuple<string, string>>();
             V4pStyles = new Dictionary<string, ASSStyle>();
             Events = new List<ASSEvent>();
@@ -48,7 +66,8 @@ namespace ASSLoader.NET
 
             // Regex defination
             var regexScriptInfoKeyValue = new Regex(@"^(?<key>[0-9a-zA-z ]+)\s*\:\s*(?<value>.+)$");
-            var unknowSectionName = string.Empty;
+            var availablePrefix = Enum.GetNames(typeof(ASSEventType));
+            var regexAvailablePrefix = new Regex(@"^(?<prefix>" + string.Join("|", availablePrefix) + @"):\s*(?<values>.+)$");
 
             for (var lineIndex = 0; lineIndex < lines.Length; lineIndex++)
             {
@@ -107,7 +126,8 @@ namespace ASSLoader.NET
 
                 switch (workingSection)
                 {
-                    // Skiplines
+                    // Unknown section processor.
+                    // Each unknown sections would keep the same while outputing.
                     case ASSSection.Unknown:
                         if (UnknownSections.ContainsKey(unknowSectionName))
                         {
@@ -119,7 +139,9 @@ namespace ASSLoader.NET
                         }
                         continue;
 
-                    // Loading ScriptInfo section.
+                    // [Script Info] Processor.
+                    // Lines starts with `!:` and `;` would be skipped because they are identified as comments.
+                    // `regexScriptInfoKeyValue` would be used for the other lines to parse the key-value pair script informations.
                     case ASSSection.ScriptInfo:
                         if (line.StartsWith("!:"))
                         {
@@ -148,7 +170,15 @@ namespace ASSLoader.NET
                         }
                         continue;
 
-                    // Loading V4 Styles+ Section.
+                    // [V4+Styles] Processor
+                    //
+                    // 1. Check the first line in this section.
+                    //    It should be started with "Format:" and followed with the headings of each column splitted by ",".
+                    //    If "Format" line was missing, the whole [V4+Styles] section would be skipped.
+                    //
+                    // 2. The remaining lines should be started with "Style:" and followed by the values of each column splitted by ",".
+                    //    * The count of the values in each line of these "Style" must as same as the count of headings in "Format".
+                    //    * `MappingToStyle` function could help to parse the list of values(not the whole string of the line) into ASSStyle object.
                     case ASSSection.V4pStyles:
                         if (V4pStyleFormat.Count == 0)
                         {
@@ -172,7 +202,7 @@ namespace ASSLoader.NET
                                     try
                                     {
                                         // mapping
-                                        var style = MappingToStyle(V4pStyleFormat, values);
+                                        var style = ASSStyle.Parse(V4pStyleFormat, values);
                                         V4pStyles[style.Name] = style;
                                     }
                                     catch (Exception exc)
@@ -195,7 +225,16 @@ namespace ASSLoader.NET
                         }
                         continue;
 
-                    // Loading Events Section.
+                    // [Events] Processor
+                    //
+                    // 1. Check the first line in this section.
+                    //    It should be started with "Format:" and followed with the headings of each column splitted by ",".
+                    //    If "Format" line was missing, the whole [Events] section would be skipped.
+                    //
+                    // 2. The remaining lines should be started with one of `ASSEventType` and ":".
+                    //    Then followed by the values of each column splitted by ",".
+                    //    * The count of the values in each line of these "Event" must as same as the count of headings in "Format".
+                    //    * `MappingToEvent` function could help to parse the list of values(not the whole string of the line) into ASSEvent object.
                     case ASSSection.Events:
                         if (EventFormat.Count == 0)
                         {
@@ -211,9 +250,7 @@ namespace ASSLoader.NET
                         }
                         else
                         {
-                            var availablePrefix = Enum.GetNames(typeof(ASSEventType));
-                            var regex = new Regex(@"^(?<prefix>" + string.Join("|", availablePrefix) + @"):\s*(?<values>.+)$");
-                            var match = regex.Match(line);
+                            var match = regexAvailablePrefix.Match(line);
                             if (match.Success)
                             {
                                 var values = match.Groups["values"].Value.Split(new[] { ',' }, EventFormat.Count).Select(x => x.Trim()).ToList();
@@ -222,7 +259,7 @@ namespace ASSLoader.NET
                                     try
                                     {
                                         // mapping
-                                        var evt = MappingToEvent(match.Groups["prefix"].Value, EventFormat, values);
+                                        var evt = ASSEvent.Parse(match.Groups["prefix"].Value, EventFormat, values);
                                         Events.Add(evt);
                                     }
                                     catch (Exception exc)
@@ -248,6 +285,10 @@ namespace ASSLoader.NET
             }
         }
 
+        /// <summary>
+        /// Convert ASSSubtitle object into ".ass" text formatted string.
+        /// </summary>
+        /// <returns>ASS formatted string.</returns>
         public string Stringify()
         {
             var sb = new StringBuilder();
@@ -258,7 +299,7 @@ namespace ASSLoader.NET
             {
                 if (si.Value.Item1.Equals("comment"))
                 {
-                    sb.AppendLine($";" + si.Value.Item2);
+                    sb.AppendLine($";{si.Value.Item2}");
                 }
                 if (si.Value.Item1.Equals("key-value"))
                 {
@@ -280,7 +321,8 @@ namespace ASSLoader.NET
             sb.AppendLine($"Format: {string.Join(", ", V4pStyleFormat)}");
             foreach (var s in V4pStyles)
             {
-                sb.AppendLine(FormatStyle(V4pStyleFormat, s.Value));
+                var style = s.Value;
+                sb.AppendLine(style.Stringify(V4pStyleFormat));
             }
             sb.AppendLine();
 
@@ -289,13 +331,18 @@ namespace ASSLoader.NET
             sb.AppendLine($"Format: {string.Join(", ", EventFormat)}");
             foreach (var ev in Events)
             {
-                sb.AppendLine(FormatEvent(EventFormat, ev));
+                sb.AppendLine(ev.Stringify(EventFormat));
             }
             sb.AppendLine();
 
             return sb.ToString();
         }
 
+        /// <summary>
+        /// Convert ASSSubtitle object into ".ass" text formatted string, and then save into a file.
+        /// </summary>
+        /// <param name="path">The path of the ".ass" file.</param>
+        /// <param name="enc">The encoding used to process the file. Default encoding is 'utf-8'(no BOM).</param>
         public void Save(string path, Encoding enc = null)
         {
             if (enc == null)
@@ -307,461 +354,6 @@ namespace ASSLoader.NET
                 File.WriteAllText(path, Stringify(), enc);
             }
         }
-
-        private static ASSEvent MappingToEvent(string prefix, IList<string> eventFormat, IList<string> values)
-        {
-            var eventType = (ASSEventType)Enum.Parse(typeof(ASSEventType), prefix);
-            var evt = new ASSEvent();
-            evt.Type = eventType;
-            for (var i = 0; i < eventFormat.Count; i++)
-            {
-                var field = eventFormat[i];
-                var value = values[i];
-                switch (field.Trim())
-                {
-                    case "Layer": evt.Layer = Convert.ToInt32(value); continue;
-                    case "Start": evt.Start = new ASSEventTime(value); continue;
-                    case "End": evt.End = new ASSEventTime(value); continue;
-                    case "Style": evt.Style = value; continue;
-                    case "Name": evt.Name = value; continue;
-                    case "MarginL": evt.MarginL = Convert.ToInt32(value); continue;
-                    case "MarginR": evt.MarginR = Convert.ToInt32(value); continue;
-                    case "MarginV": evt.MarginV = Convert.ToInt32(value); continue;
-                    case "Effect": evt.Effect = value; continue;
-                    case "Text": evt.Text = value; continue;
-                    default:
-                        Trace.TraceWarning("MAPPING ERROR: Unknown field - " + field);
-                        continue;
-                }
-            }
-            return evt;
-        }
-
-        private static string FormatEvent(IList<string> eventFormat, ASSEvent evt, string spliter = ",")
-        {
-            var sb = new StringBuilder();
-            sb.Append(evt.Type.ToString() + ": ");
-            for (var i = 0; i < eventFormat.Count; i++)
-            {
-                var field = eventFormat[i];
-                switch (field.Trim())
-                {
-                    case "Layer": sb.Append(evt.Layer); break;
-                    case "Start": sb.Append(evt.Start); break;
-                    case "End": sb.Append(evt.End); break;
-                    case "Style": sb.Append(evt.Style); break;
-                    case "Name": sb.Append(evt.Name); break;
-                    case "MarginL": sb.Append(evt.MarginL); break;
-                    case "MarginR": sb.Append(evt.MarginR); break;
-                    case "MarginV": sb.Append(evt.MarginV); break;
-                    case "Effect": sb.Append(evt.Effect); break;
-                    case "Text": sb.Append(evt.Text); break;
-                    default:
-                        Trace.TraceWarning("MAPPING ERROR: Unknown field - " + field);
-                        break;
-                }
-                if (i != eventFormat.Count - 1)
-                {
-                    sb.Append(spliter);
-                }
-            }
-            return sb.ToString();
-        }
-
-        private static ASSStyle MappingToStyle(IList<string> v4pStyleFormat, IList<string> values)
-        {
-            var style = new ASSStyle();
-            for (var i = 0; i < v4pStyleFormat.Count; i++)
-            {
-                var field = v4pStyleFormat[i];
-                var value = values[i];
-                switch (field.Trim())
-                {
-                    case "Name": style.Name = value; continue;
-                    case "Fontname": style.Fontname = value; continue;
-                    case "Fontsize": style.Fontsize = Convert.ToDouble(value); continue;
-                    case "PrimaryColour": style.PrimaryColour = value; continue;
-                    case "SecondaryColour": style.SecondaryColour = value; continue;
-                    case "OutlineColour": style.OutlineColour = value; continue;
-                    case "BackColour": style.BackColour = value; continue;
-                    case "Bold": style.Bold = Convert.ToInt16(value) == -1; continue;
-                    case "Italic": style.Italic = Convert.ToInt16(value) == -1; continue;
-                    case "Underline": style.Underline = Convert.ToInt16(value) == -1; continue;
-                    case "StrikeOut": style.StrikeOut = Convert.ToInt16(value) == -1; continue;
-                    case "ScaleX": style.ScaleX = Convert.ToDouble(value); continue;
-                    case "ScaleY": style.ScaleY = Convert.ToDouble(value); continue;
-                    case "Spacing": style.Spacing = Convert.ToInt32(value); continue;
-                    case "Angle": style.Angle = Convert.ToDouble(value); continue;
-                    case "BorderStyle": style.BorderStyle = (V4pStyleBorderStyle)Convert.ToInt16(value); continue;
-                    case "Outline": style.Outline = Convert.ToInt32(value); continue;
-                    case "Shadow": style.Shadow = Convert.ToInt32(value); continue;
-                    case "Alignment": style.Alignment = (V4pStyleAlignment)Convert.ToInt16(value); continue;
-                    case "MarginL": style.MarginL = Convert.ToInt32(value); continue;
-                    case "MarginR": style.MarginR = Convert.ToInt32(value); continue;
-                    case "MarginV": style.MarginV = Convert.ToInt32(value); continue;
-                    case "Encoding": style.Encoding = Convert.ToInt32(value); continue;
-                    default:
-                        Trace.TraceWarning("MAPPING ERROR: Unknown field - " + field);
-                        continue;
-                }
-            }
-            return style;
-        }
-
-        private static string FormatStyle(IList<string> v4pStyleFormat, ASSStyle style, string spliter = ",")
-        {
-            var sb = new StringBuilder();
-            sb.Append("Style: ");
-            for (var i = 0; i < v4pStyleFormat.Count; i++)
-            {
-                var field = v4pStyleFormat[i];
-                switch (field.Trim())
-                {
-                    case "Name": sb.Append(style.Name); break;
-                    case "Fontname": sb.Append(style.Fontname); break;
-                    case "Fontsize": sb.Append(style.Fontsize); break;
-                    case "PrimaryColour": sb.Append(style.PrimaryColour); break;
-                    case "SecondaryColour": sb.Append(style.SecondaryColour); break;
-                    case "OutlineColour": sb.Append(style.OutlineColour); break;
-                    case "BackColour": sb.Append(style.BackColour); break;
-                    case "Bold": sb.Append(style.Bold ? "-1" : "0"); break;
-                    case "Italic": sb.Append(style.Italic ? "-1" : "0"); break;
-                    case "Underline": sb.Append(style.Underline ? "-1" : "0"); break;
-                    case "StrikeOut": sb.Append(style.StrikeOut ? "-1" : "0"); break;
-                    case "ScaleX": sb.Append(style.ScaleX); break;
-                    case "ScaleY": sb.Append(style.ScaleY); break;
-                    case "Spacing": sb.Append(style.Spacing); break;
-                    case "Angle": sb.Append(style.Angle); break;
-                    case "BorderStyle": sb.Append((int)style.BorderStyle); break;
-                    case "Outline": sb.Append(style.Outline); break;
-                    case "Shadow": sb.Append(style.Shadow); break;
-                    case "Alignment": sb.Append((int)style.Alignment); break;
-                    case "MarginL": sb.Append(style.MarginL); break;
-                    case "MarginR": sb.Append(style.MarginR); break;
-                    case "MarginV": sb.Append(style.MarginV); break;
-                    case "Encoding": sb.Append(style.Encoding); break;
-                    default:
-                        Trace.TraceWarning("MAPPING ERROR: Unknown field - " + field);
-                        break;
-                }
-                if (i != v4pStyleFormat.Count - 1)
-                {
-                    sb.Append(spliter);
-                }
-            }
-            return sb.ToString();
-        }
-
     } // class ASSSubtitle
 
-    public class ASSStyle
-    {
-        /// <summary>
-        /// The name of the Style. Case sensitive. Cannot include commas.
-        /// </summary>
-        public string Name { get; set; }
-
-        /// <summary>
-        /// The fontname as used by Windows. Case-sensitive.
-        /// </summary>
-        public string Fontname { get; set; }
-
-        /// <summary>
-        /// The fontsize.
-        /// </summary>
-        public double Fontsize { get; set; }
-
-        public string PrimaryColour { get; set; }
-
-        public string SecondaryColour { get; set; }
-
-        public string OutlineColour { get; set; }
-
-        public string BackColour { get; set; }
-
-        public bool Bold { get; set; }
-
-        public bool Italic { get; set; }
-
-        public bool Underline { get; set; }
-
-        public bool StrikeOut { get; set; }
-
-        public double ScaleX { get; set; }
-
-        public double ScaleY { get; set; }
-
-        public int Spacing { get; set; }
-
-        public double Angle { get; set; }
-
-        public V4pStyleBorderStyle BorderStyle { get; set; }
-
-        public int Outline { get; set; }
-
-        public int Shadow { get; set; }
-
-        public V4pStyleAlignment Alignment { get; set; }
-
-        public int MarginL { get; set; }
-
-        public int MarginR { get; set; }
-
-        public int MarginV { get; set; }
-
-        public int AlphaLevel { get; set; }
-
-        public int Encoding { get; set; }
-
-        public override string ToString()
-        {
-            if (ASSSubtitle.showAbstract)
-            {
-                return $"{Name}: {Fontname},{Fontsize}{(Bold ? "B" : "")}{(Italic ? "I" : "")}{(Underline ? "U" : "")}{(StrikeOut ? "S" : "")}";
-            }
-            else
-            {
-                base.ToString();
-            }
-        }
-    }
-
-    public class ASSEvent : ICloneable
-    {
-        public ASSEventType Type { get; set; }
-
-        public int Layer { get; set; }
-
-        public ASSEventTime Start { get; set; }
-
-        public ASSEventTime End { get; set; }
-
-        public string Style { get; set; }
-
-        public string Name { get; set; }
-
-        public int MarginL { get; set; }
-
-        public int MarginR { get; set; }
-
-        public int MarginV { get; set; }
-
-        public string Effect { get; set; }
-
-        public string Text { get; set; }
-
-        public object Clone()
-        {
-            var newInst = new ASSEvent();
-            newInst.Type = this.Type;
-            newInst.Layer = this.Layer;
-            newInst.Start = new ASSEventTime(this.Start.ToString());
-            newInst.End = new ASSEventTime(this.End.ToString());
-            newInst.Style = this.Style;
-            newInst.Name = this.Name;
-            newInst.MarginL = this.MarginL;
-            newInst.MarginR = this.MarginR;
-            newInst.MarginV = this.MarginV;
-            newInst.Effect = this.Effect;
-            newInst.Text = this.Text;
-            return newInst;
-        }
-
-        public ASSEvent()
-        {
-            this.Type = ASSEventType.Dialogue;
-            this.Layer = 0;
-            this.Start = new ASSEventTime(0, 0, 0, 0);
-            this.End = new ASSEventTime(0, 0, 0, 0);
-            this.Style = "Default";
-            this.Name = string.Empty;
-            this.MarginL = 0;
-            this.MarginR = 0;
-            this.MarginV = 0;
-            this.Effect = string.Empty;
-            this.Text = string.Empty;
-        }
-
-        public override string ToString()
-        {
-            if (ASSSubtitle.showAbstract)
-            {
-                return $"{Start} - {End} | {Type}:{Name}:{Text}";
-            }
-            else
-            {
-                base.ToString();
-            }
-        }
-    }
-
-    public class ASSEventTime : IComparable
-    {
-        public int Hour { get; set; }
-
-        public int Minute { get; set; }
-
-        public int Second { get; set; }
-
-        public int Millisecond { get; set; }
-
-        public ASSEventTime(string assTime)
-        {
-            var parts = assTime.Split(':', '.');
-            var msIndex = parts.Length - 1;
-            var secIndex = parts.Length - 2;
-            var minIndex = parts.Length - 3;
-            var hourIndex = parts.Length - 4;
-            this.Hour = hourIndex > 0 ? Convert.ToInt32(parts[hourIndex]) : 0;
-            this.Minute = minIndex > 0 ? Convert.ToInt32(parts[minIndex]) : 0;
-            this.Second = secIndex > 0 ? Convert.ToInt32(parts[secIndex]) : 0;
-            this.Millisecond = msIndex > 0 ? Convert.ToInt32((parts[msIndex] + "000").Substring(0, 3)) : 0;
-        }
-
-        public ASSEventTime(int hour, int minute, int second, int millisecond)
-        {
-            this.Hour = hour;
-            this.Minute = minute;
-            this.Second = second;
-            this.Millisecond = millisecond;
-        }
-
-        public long TotalMilliseconds()
-        {
-            return this.Hour * 3600000 + this.Minute * 60000 + this.Second * 1000 + this.Millisecond;
-        }
-
-        public static explicit operator ASSEventTime(TimeSpan ts)
-        {
-            return new ASSEventTime(ts.Hours, ts.Minutes, ts.Seconds, ts.Milliseconds);
-        }
-
-        public static explicit operator TimeSpan(ASSEventTime time)
-        {
-            return new TimeSpan(0, time.Hour, time.Minute, time.Second, time.Millisecond);
-        }
-
-        public static ASSEventTime operator +(ASSEventTime aet, double num)
-        {
-            var target = new ASSEventTime(aet.ToString());
-            var ms = Convert.ToInt32(Math.Floor(num * 1000));
-            target.Millisecond = target.Millisecond + ms;
-            if (target.Millisecond > 1000)
-            {
-                target.Second += target.Millisecond / 1000;
-                target.Millisecond = target.Millisecond % 1000;
-            }
-            if (target.Second > 60)
-            {
-                target.Minute += target.Second / 60;
-                target.Second = target.Second % 60;
-            }
-            if(target.Minute > 60)
-            {
-                target.Hour += target.Minute / 60;
-                target.Minute = target.Minute % 60;
-            }
-
-            return target;
-        }
-
-        public static ASSEventTime operator -(ASSEventTime aet, double num)
-        {
-            var ms = Convert.ToInt32(Math.Floor(num * 1000));
-            var target = new ASSEventTime(aet.ToString());
-            target.Millisecond = aet.Millisecond - ms;
-            if (target.Millisecond < 0)
-            {
-                target.Millisecond += 1000;
-                target.Second -= 1;
-            }
-            if (target.Second < 0)
-            {
-                target.Second += 60;
-                target.Minute -= 1;
-            }
-            if (target.Minute < 0)
-            {
-                target.Minute += 60;
-                target.Hour -= 1;
-            }
-            return target;
-        }
-
-        public override bool Equals(object obj)
-        {
-            return CompareTo(obj) == 0;
-        }
-
-        public override int GetHashCode()
-        {
-            return (int)this.TotalMilliseconds();
-        }
-        public override string ToString()
-        {
-            return this.Hour.ToString().Substring(0, 1) + ":"
-                   + this.Minute.ToString().PadLeft(2, '0') + ":"
-                   + this.Second.ToString().PadLeft(2, '0') + "."
-                   + this.Millisecond.ToString().PadLeft(3, '0').Substring(0, 2);
-        }
-
-        public int CompareTo(object obj)
-        {
-            var target = obj as ASSEventTime;
-            if (target == null)
-            {
-                target = new ASSEventTime(obj.ToString());
-            }
-            return (int)(this.TotalMilliseconds() - target.TotalMilliseconds());
-        }
-    }
-
-    public class ASSEmbeddedFont
-    {
-
-    }
-
-    public class ASSEmbeddedGraphics
-    {
-
-    }
-
-    public enum V4pStyleBorderStyle
-    {
-        BorderAndShadow = 1,
-        ColorBackground = 3
-    }
-
-    public enum V4pStyleAlignment
-    {
-        SubLF = 1,
-        SubCT = 2,
-        SubRT = 3,
-        MidLF = 4,
-        MidCT = 5,
-        MidRT = 6,
-        TopLF = 7,
-        TopCT = 8,
-        TopRT = 9
-    }
-
-    public enum ASSEventType
-    {
-        Dialogue,
-        Comment,
-        Picture,
-        Sound,
-        Movie,
-        Command
-    }
-
-    public enum ASSSection
-    {
-        Unknown,
-        ScriptInfo,
-        V4pStyles,
-        Events,
-        Fonts,
-        Graphics
-    }
 }
